creates user dashboard component to stop re-render of the search bar
