+++ apps/gp2-frontend/src/users/Routes.tsx
@@ -1,40 +1,28 @@
-import { Frame } from '@asap-hub/frontend-utils';
-import { UsersPage } from '@asap-hub/gp2-components';
 import { NotFoundPage } from '@asap-hub/react-components';
 import { gp2 } from '@asap-hub/routing';
 import { lazy, useEffect } from 'react';
 import { Route, Switch, useRouteMatch } from 'react-router-dom';
+import UserDashboard from './UserDashboard';
 
-const loadUserList = () =>
-  import(/* webpackChunkName: "user-list" */ './UserList');
 const loadUserDetail = () =>
   import(/* webpackChunkName: "user-detail" */ './UserDetail');
 
-const UserList = lazy(loadUserList);
 const UserDetail = lazy(loadUserDetail);
 
 const { users } = gp2;
 const Routes: React.FC<Record<string, never>> = () => {
   useEffect(() => {
-    loadUserList().then(loadUserDetail);
+    loadUserDetail();
   }, []);
   const { path } = useRouteMatch();
 
   return (
     <Switch>
       <Route exact path={path}>
-        <UsersPage>
-          <Frame title="Users">
-            <UserList />
-          </Frame>
-        </UsersPage>
+        <UserDashboard />
       </Route>
       <Route exact path={users({}).filters({}).$}>
-        <UsersPage>
-          <Frame title="Users">
-            <UserList displayFilters />
-          </Frame>
-        </UsersPage>
+        <UserDashboard displayFilters />
       </Route>
       <Route path={path + users({}).user.template}>
         <UserDetail />
