+++ apps/storybook/src/gp2-components/UsersPage.stories.tsx
@@ -1,4 +1,6 @@
 import { Layout, UsersPage } from '@asap-hub/gp2-components';
+import { action } from '@storybook/addon-actions';
+import { select, text } from '@storybook/addon-knobs';
 import { NoPaddingDecorator } from '../layout';
 
 export default {
@@ -9,6 +11,24 @@ export default {
 
 export const Normal = () => (
   <Layout>
-    <UsersPage>Page content</UsersPage>
+    <UsersPage
+      onSearchQueryChange={() => action('search')}
+      searchQuery={text('Search Query', '')}
+      isAdministrator={
+        (select<boolean>, 'Is Administrator', { Yes: true, No: false }, false)
+      }
+      onFiltersClick={() => action('filters')}
+      onExportClick={() => action('export')}
+      changeLocation={() => action('location')}
+      updateFilters={() => action('update filters')}
+      filters={{}}
+      projects={[]}
+      workingGroups={[]}
+      displayFilters={
+        (select<boolean>, 'Display Filters', { Yes: true, No: false }, false)
+      }
+    >
+      Page content
+    </UsersPage>
   </Layout>
 );
