+++ packages/gp2-components/src/organisms/UsersPageBody.tsx
@@ -2,14 +2,12 @@ import { gp2 } from '@asap-hub/model';
 import { PageControls, pixels } from '@asap-hub/react-components';
 import { css } from '@emotion/react';
 import { ComponentProps } from 'react';
-import FilterSearchExport from './FilterSearchExport';
 import UserCard from './UserCard';
 
 const { rem } = pixels;
 type UsersPageBodyProps = {
   users: gp2.ListUserResponse;
-} & ComponentProps<typeof PageControls> &
-  ComponentProps<typeof FilterSearchExport>;
+} & ComponentProps<typeof PageControls>;
 
 const containerStyles = css({
   display: 'flex',
@@ -20,21 +18,9 @@ const containerStyles = css({
 
 const UsersPageBody: React.FC<UsersPageBodyProps> = ({
   users,
-  onFiltersClick,
-  onExportClick,
-  isAdministrator,
-  searchQuery,
-  onSearchQueryChange,
   ...pageProps
 }) => (
   <article css={containerStyles}>
-    <FilterSearchExport
-      searchQuery={searchQuery}
-      onSearchQueryChange={onSearchQueryChange}
-      onFiltersClick={onFiltersClick}
-      onExportClick={onExportClick}
-      isAdministrator={isAdministrator}
-    />
     {users.items.map((user) => (
       <UserCard key={user.id} {...user} />
     ))}
